requirements.txt
pandas==2.2.3
scikit-learn==1.7.2
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
datasets[audio]<3
matplotlib
Pillow
lpips 
scikit-image 
tqdm